Fix Hw Decoder inspector field not dynamic
This commit will change InspectorWindowController.updateInfo to treat
the Hw Decoder field as being dynamic. At any point in time while the
video is playing hardware decoding may fail causing a fall back to
software decoding. The inspector window needs to treat this field as
dynamic to insure the type of decoding shown is current.